+++ sqlglot/tokens.py
@@ -85,6 +85,7 @@ class TokenType(AutoName):
     UTINYINT = auto()
     SMALLINT = auto()
     USMALLINT = auto()
+    MEDIUMINT = auto()
     INT = auto()
     UINT = auto()
     BIGINT = auto()
@@ -141,6 +142,7 @@ class TokenType(AutoName):
     SMALLSERIAL = auto()
     BIGSERIAL = auto()
     XML = auto()
+    YEAR = auto()
     UNIQUEIDENTIFIER = auto()
     USERDEFINED = auto()
     MONEY = auto()
@@ -182,6 +184,7 @@ class TokenType(AutoName):
     COMMAND = auto()
     COMMENT = auto()
     COMMIT = auto()
+    CONNECT_BY = auto()
     CONSTRAINT = auto()
     CREATE = auto()
     CROSS = auto()
@@ -300,6 +303,7 @@ class TokenType(AutoName):
     SIMILAR_TO = auto()
     SOME = auto()
     SORT_BY = auto()
+    START_WITH = auto()
     STRUCT = auto()
     TABLE_SAMPLE = auto()
     TEMPORARY = auto()
@@ -532,6 +536,7 @@ class Tokenizer(metaclass=_Tokenizer):
         "COLLATE": TokenType.COLLATE,
         "COLUMN": TokenType.COLUMN,
         "COMMIT": TokenType.COMMIT,
+        "CONNECT BY": TokenType.CONNECT_BY,
         "CONSTRAINT": TokenType.CONSTRAINT,
         "CREATE": TokenType.CREATE,
         "CROSS": TokenType.CROSS,
@@ -638,6 +643,7 @@ class Tokenizer(metaclass=_Tokenizer):
         "SIMILAR TO": TokenType.SIMILAR_TO,
         "SOME": TokenType.SOME,
         "SORT BY": TokenType.SORT_BY,
+        "START WITH": TokenType.START_WITH,
         "TABLE": TokenType.TABLE,
         "TABLESAMPLE": TokenType.TABLE_SAMPLE,
         "TEMP": TokenType.TEMPORARY,